render sub folders indented
